no more array size

+++ backend/include/bridge/types.h
@@ -41,7 +41,7 @@ struct RefTy {
 
     // for RArray
     std::unique_ptr<Ty> inner;
-    int64_t             size = 0;
+    int             size = 0;
 
     // for RClass
     std::string cname;
